requirements_final.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
app==0.0.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
backports.csv==1.0.7
beautifulsoup4==4.10.0
bert-score==0.3.11
bleach==4.1.0
blinker==1.4
blis==0.7.4
bookkeywords==1.0.0
boto3==1.18.44
botocore==1.21.44
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cheroot==8.5.2
CherryPy==18.6.1
click==7.1.2
colorama==0.4.4
Counter==1.0.0
cryptography==3.4.8
cycler==0.10.0
cymem==2.0.5
datasets==2.3.2
dill==0.3.5.1
docutils==0.17.1
emoji==1.5.0
emotions==1.2.7
et-xmlfile==1.1.0
fastai==2.6.3
fastcore==1.4.4
fastdownload==0.0.6
fastprogress==1.0.2
fasttext==0.9.2
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.1
Flask-CacheBuster==1.0.0
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-ReCaptcha==0.4.2
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gensim==3.8.3
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.7.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.4.0
jaraco.text==3.6.0
jeepney==0.7.1
jellyfish==0.8.8
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
keyring==23.1.0
kiwisolver==1.3.2
lime==0.2.0.1
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.4.3
more-itertools==8.9.0
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.5
mysqlclient==2.0.3
networkx==2.6.3
nltk==3.6.2
numpy==1.21.2
ohmeow-blurr==1.0.5
openpyxl==3.0.10
packaging==21.0
pandas==1.3.3
pathy==0.6.0
Pattern==3.6
pdfminer.six==20211012
pickle-mixin==1.0.2
Pillow==8.3.2
pkginfo==1.7.1
plac==1.1.3
pluggy==1.0.0
portend==3.0.0
preshed==3.0.5
py==1.10.0
pyarrow==8.0.0
pybind11==2.9.2
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==4.4.1
pytest-runner==4.4
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
rake-nltk==1.0.4
readme-renderer==29.0
regex==2021.8.28
requests==2.26.0
requests-toolbelt==0.9.1
responses==0.18.0
rfc3986==1.5.0
rouge-score==0.0.4
s3transfer==0.5.0
sacremoses==0.0.45
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
SecretStorage==3.3.1
segtok==1.5.10
sentencepiece==0.1.96
seqeval==1.2.2
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3
spacy==3.1.2
spacy-legacy==3.0.8
SQLAlchemy==1.4.37
srsly==2.4.1
tabulate==0.8.9
tempora==4.1.2
text2emotion==0.0.5
textblob==0.15.3
textgenie==0.1.9.3
thinc==8.0.10
threadpoolctl==2.2.0
tifffile==2021.8.30
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.2
transformers==4.19.2
twine==3.4.2
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
webencodings==0.5.1
Werkzeug==2.0.1
wordcloud==1.8.2.2
xlrd==2.0.1
xxhash==3.0.0
yake==0.4.8
yarl==1.7.2
zc.lockfile==2.0
zipp==3.5.0
